refactor(settings): split GeneralSection into per-group components
Break the 445-line GeneralSection monolith into four focused components
under setting/general/ — StartupSettings, LanguageSettings,
TelemetrySettings, and DiagnosticsSettings — plus a shared useSavingState
hook that gives every save handler the same try/flag/catch shape.

Each section now owns its own saving flag, so an in-flight save only
disables that section's controls. The debug-mode confirm/restart flow is
modeled as an explicit closed|confirming|restarting state machine. No
behavior, UI, or i18n keys change; GeneralSection stays the default export.
